rename: 패키지명 변경
- pieces 는 개별 체스말들을 모두 아울러서 부르는 것 같음
