+++ src/main/java/com/kyu/chess/piece/Piece.java
@@ -1,4 +1,4 @@
-package com.kyu.chess.pieces;
+package com.kyu.chess.piece;
 
 public class Piece {
     private PieceStatus pieceStatus;
