+++ src/test/java/com/kyu/chess/piece/PieceTest.java
@@ -1,4 +1,4 @@
-package com.kyu.chess.pieces;
+package com.kyu.chess.piece;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
